Extend and clean up draft
Consolidate some of my findings from programming the library.

Remove doubt about including coordinates in `Entity`. We need the
coordinates if we represent the data in a sequential manner as we do
right now. And even in a hierarchical presentation it would be much
nicer not to need to derive them from the data structure.