AspectsImp: Corrections, especially prerequisites
Correct some things in the Aspects guide. The main thing is that
prerequisites are no longer specified as a set, but as a function. The
primitive approach with the set doesn't work. because it doesn't support
ORs. – "this Aspect OR that Aspect" – this might be solved with data,
but is likely to be more clumsy. If specification of prerequisites
through data is desired, it may be added later. – Functions are rather
flexible.

Apart from that, small changes:

 - Rename <aspect name>-info to <aspect name>-def. We're defining an
   aspect and this is the part of the definition that can be encoded, so
   it kind of makes more sense.

 - Change that part about dependency resolution. – It doesn't make sense
   since we can't take any action when a prerequisite is not fulfilled
   or something like that. Prerequisites are information and they can
   rely on other information.

 - Change <name validator> to <name predicate> in order to make it
   consistent with the <prerequisites predicate>.

 - Change <…-…> to <… …>. < and > are delimitation enough and this way
   it's easier to read.