BarsImp: Add possibility to provide schema
While implementing the Bars stuff, I found it more helpful to provide a
Schema directly instead of requiring the Bar type definer to use the
Schema in a predicate. With a predicate, the only information you can
get is true or false. With a Schema, the user can find out what part of
the data structure was malformed. Update the spec to include the schema
option.